+++ src/Pages/Dashboard.jsx
@@ -1,0 +1,44 @@
+import * as React from 'react'
+import Box from '@mui/material/Box'
+import SideNav from '../Components/SideNav'
+import AppBar from '@mui/material/AppBar'
+import Toolbar from '@mui/material/Toolbar'
+import Typography from '@mui/material/Typography'
+import { AdapterDayjs } from '@mui/x-date-pickers/AdapterDayjs'
+import { LocalizationProvider } from '@mui/x-date-pickers/LocalizationProvider'
+import { StaticDatePicker } from '@mui/x-date-pickers/StaticDatePicker'
+
+
+const drawerWidth = 240;
+
+export default function Dashboard(){
+
+  return(
+
+    <>
+    <Box sx={{display: "flex", marginTop:10}}>
+      <SideNav/>
+      <AppBar
+        position="fixed"
+        sx={{
+          width: { sm: `calc(100% - ${drawerWidth}px)` },
+          ml: { sm: `${drawerWidth}px` },
+        }}
+      >
+        <Toolbar>
+          <Typography variant="h6" noWrap component="div">
+            Dashboard
+          </Typography>
+        </Toolbar>
+    </AppBar>
+    <LocalizationProvider dateAdapter={AdapterDayjs}>
+      <StaticDatePicker orientation="landscape" />
+    </LocalizationProvider>
+    </Box>
+    </>
+
+
+  )
+
+
+}
