--- src/Pages/EmployeeList.js
@@ -1,10 +1,0 @@
-import React from 'react'
-import SideNav from '../Components/SideNav'
-
-export default function EmployeeList() {
-  return (
-    <>
-    <SideNav/>
-    </>
-  )
-}
