+++ src/Pages/EmployeeList.jsx
@@ -1,0 +1,32 @@
+import * as React from 'react'
+import SideNav from '../Components/SideNav'
+import Box from '@mui/material/Box'
+import AppBar from '@mui/material/AppBar'
+import Toolbar from '@mui/material/Toolbar'
+import Typography from '@mui/material/Typography'
+
+
+const drawerWidth = 240;
+
+export default function EmployeeList() {
+  return (
+    <>
+    <Box sx={{display: "flex" }}>
+    <SideNav/>
+    <AppBar
+        position="fixed"
+        sx={{
+          width: { sm: `calc(100% - ${drawerWidth}px)` },
+          ml: { sm: `${drawerWidth}px` },
+        }}
+      >
+        <Toolbar>
+          <Typography variant="h6" noWrap component="div">
+            Employee List
+          </Typography>
+        </Toolbar>
+    </AppBar>
+    </Box>
+    </>
+  )
+}
